+++ NEWS.md
@@ -235,4 +235,5 @@
 * Guess results now reverse the normalization as well.
 * Stratified poisson models now report results relative to the total observed strata, and list the number of non-zero strata and rows.
 * Fixed bug in likelihood-based boundaries in models with constant parameters
+* formula missing term 0 will automatically assign an intercept covariate to the loglinear subterm of term 0. By default, this intercept is not constant.